% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_counts.R
\name{plot_counts}
\alias{plot_counts}
\title{Plot count data as}
\usage{
plot_counts(
  data,
  x,
  by,
  xlab = "",
  ylab = "",
  bylab = "",
  nlab = "",
  flip = FALSE
)
}
\arguments{
\item{data}{a dataframe}

\item{x}{categorical variable to count}

\item{flip}{}
}
\description{
Plot count data as
}
